# Conflicts:
#	contingent/application/contingent-area/src/main/java/moscow/ptnl/contingent/area/transform/v4/CodeNameTypeMapperV4.java
#	contingent/application/contingent-area/src/main/java/moscow/ptnl/contingent/area/ws/v4/AreaServiceImpl.java
#	contingent/application/contingent-area/src/test/java/area/service/MockConfiguration.java

+++ contingent/application/contingent-nsi-domain/src/main/java/moscow/ptnl/contingent/nsi/domain/area/AreaType.java
@@ -133,7 +133,7 @@ public class AreaType extends CodeName implements Serializable, NsiExternalEntit
     private String source;
 
     @Column(name = "AREA_TYPE_CATEGORY")
-    private Long areaTypeCategory;
+    private Integer areaTypeCategory;
 
     @OneToMany(fetch = FetchType.LAZY, mappedBy = "areaType")
     private Set<AreaTypeSpecializations> areaTypeSpecializations;
@@ -351,11 +351,11 @@ public class AreaType extends CodeName implements Serializable, NsiExternalEntit
         this.areaTypeProfiles = areaTypeProfiles;
     }
 
-    public Long getAreaTypeCategory() {
+    public Integer getAreaTypeCategory() {
         return areaTypeCategory;
     }
 
-    public void setAreaTypeCategory(Long areaTypeCategory) {
+    public void setAreaTypeCategory(Integer areaTypeCategory) {
         this.areaTypeCategory = areaTypeCategory;
     }
